修复 FILE.seek操作完成后未准备下一个contentReader的问题

+++ SJMediaCacheServer/Core/Resource/HLS/HLSAsset.m
@@ -134,7 +134,7 @@
         }
     });
 
-    NSUInteger totalLength = response.expectedContentLength;
+    NSUInteger totalLength = (NSUInteger)response.expectedContentLength;
     MCSAssetContent *content = nil;
     NSString *TsName = [MCSURLRecognizer.shared nameWithUrl:URL.absoluteString extension:HLSFileExtensionTS];
     NSString *filename = [MCSFileManager HLS_createContentFileInAsset:self.name tsName:TsName tsTotalLength:totalLength];
